requirements.txt
appdirs==1.4.4
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
idna==2.10
importlib-metadata==2.1.1
importlib-resources==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.5.2
MarkupSafe==1.1.1
MouseInfo==0.1.3
numpy==1.16.6
pathlib2==2.3.5
Pillow==6.2.2
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==5.3
pyobjc-core==5.3
pyobjc-framework-Accounts==5.3
pyobjc-framework-AddressBook==5.3
pyobjc-framework-AdSupport==5.3
pyobjc-framework-AppleScriptKit==5.3
pyobjc-framework-AppleScriptObjC==5.3
pyobjc-framework-ApplicationServices==5.3
pyobjc-framework-Automator==5.3
pyobjc-framework-AVFoundation==5.3
pyobjc-framework-AVKit==5.3
pyobjc-framework-BusinessChat==5.3
pyobjc-framework-CalendarStore==5.3
pyobjc-framework-CFNetwork==5.3
pyobjc-framework-CloudKit==5.3
pyobjc-framework-Cocoa==5.3
pyobjc-framework-Collaboration==5.3
pyobjc-framework-ColorSync==5.3
pyobjc-framework-Contacts==5.3
pyobjc-framework-ContactsUI==5.3
pyobjc-framework-CoreAudio==5.3
pyobjc-framework-CoreAudioKit==5.3
pyobjc-framework-CoreBluetooth==5.3
pyobjc-framework-CoreData==5.3
pyobjc-framework-CoreLocation==5.3
pyobjc-framework-CoreMedia==5.3
pyobjc-framework-CoreMediaIO==5.3
pyobjc-framework-CoreML==5.3
pyobjc-framework-CoreServices==5.3
pyobjc-framework-CoreSpotlight==5.3
pyobjc-framework-CoreText==5.3
pyobjc-framework-CoreWLAN==5.3
pyobjc-framework-CryptoTokenKit==5.3
pyobjc-framework-DictionaryServices==5.3
pyobjc-framework-DiscRecording==5.3
pyobjc-framework-DiscRecordingUI==5.3
pyobjc-framework-DiskArbitration==5.3
pyobjc-framework-DVDPlayback==5.3
pyobjc-framework-EventKit==5.3
pyobjc-framework-ExceptionHandling==5.3
pyobjc-framework-ExternalAccessory==5.3
pyobjc-framework-FinderSync==5.3
pyobjc-framework-FSEvents==5.3
pyobjc-framework-GameCenter==5.3
pyobjc-framework-GameController==5.3
pyobjc-framework-GameKit==5.3
pyobjc-framework-GameplayKit==5.3
pyobjc-framework-ImageCaptureCore==5.3
pyobjc-framework-IMServicePlugIn==5.3
pyobjc-framework-InputMethodKit==5.3
pyobjc-framework-InstallerPlugins==5.3
pyobjc-framework-InstantMessage==5.3
pyobjc-framework-Intents==5.3
pyobjc-framework-IOSurface==5.3
pyobjc-framework-iTunesLibrary==5.3
pyobjc-framework-LatentSemanticMapping==5.3
pyobjc-framework-LaunchServices==5.3
pyobjc-framework-libdispatch==5.3
pyobjc-framework-LocalAuthentication==5.3
pyobjc-framework-MapKit==5.3
pyobjc-framework-MediaAccessibility==5.3
pyobjc-framework-MediaLibrary==5.3
pyobjc-framework-MediaPlayer==5.3
pyobjc-framework-MediaToolbox==5.3
pyobjc-framework-ModelIO==5.3
pyobjc-framework-MultipeerConnectivity==5.3
pyobjc-framework-NaturalLanguage==5.3
pyobjc-framework-NetFS==5.3
pyobjc-framework-Network==5.3
pyobjc-framework-NetworkExtension==5.3
pyobjc-framework-NotificationCenter==5.3
pyobjc-framework-OpenDirectory==5.3
pyobjc-framework-OSAKit==5.3
pyobjc-framework-Photos==5.3
pyobjc-framework-PhotosUI==5.3
pyobjc-framework-PreferencePanes==5.3
pyobjc-framework-PubSub==5.3
pyobjc-framework-QTKit==5.3
pyobjc-framework-Quartz==5.3
pyobjc-framework-SafariServices==5.3
pyobjc-framework-SceneKit==5.3
pyobjc-framework-ScreenSaver==5.3
pyobjc-framework-ScriptingBridge==5.3
pyobjc-framework-SearchKit==5.3
pyobjc-framework-Security==5.3
pyobjc-framework-SecurityFoundation==5.3
pyobjc-framework-SecurityInterface==5.3
pyobjc-framework-ServiceManagement==5.3
pyobjc-framework-Social==5.3
pyobjc-framework-SpriteKit==5.3
pyobjc-framework-StoreKit==5.3
pyobjc-framework-SyncServices==5.3
pyobjc-framework-SystemConfiguration==5.3
pyobjc-framework-UserNotifications==5.3
pyobjc-framework-VideoSubscriberAccount==5.3
pyobjc-framework-VideoToolbox==5.3
pyobjc-framework-Vision==5.3
pyobjc-framework-WebKit==5.3
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
PyTweening==1.0.3
PyYAML==5.4.1
requests==2.25.1
rubicon-objc==0.2.10
scandir==1.10.0
singledispatch==3.4.0.3
six==1.15.0
soupsieve==1.9.6
typing==3.7.4.3
urllib3==1.26.4
virtualenv==20.2.1
Werkzeug==1.0.1
zipp==1.2.0